+++ src/main/java/kitchenpos/menus/tobe/domain/PurgomalumClient.java
@@ -1,0 +1,5 @@
+package kitchenpos.menus.tobe.domain;
+
+public interface PurgomalumClient {
+    boolean containsProfanity(String text);
+}
