+++ src/test/java/kitchenpos/menus/application/MenuServiceTest.java
@@ -8,10 +8,10 @@ import kitchenpos.menus.tobe.domain.Menu;
 import kitchenpos.menus.domain.MenuGroupRepository;
 import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.menus.tobe.domain.Product;
-import kitchenpos.products.application.FakePurgomalumClient;
+import kitchenpos.clients.FakePurgomalumClient;
 import kitchenpos.products.application.InMemoryProductRepository;
 import kitchenpos.products.tobe.domain.ProductRepository;
-import kitchenpos.products.tobe.domain.PurgomalumClient;
+import kitchenpos.menus.tobe.domain.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
