+++ src/test/java/kitchenpos/menus/application/ProductPriceChangeListenerTest.java
@@ -5,9 +5,12 @@ import kitchenpos.menus.domain.MenuGroupRepository;
 import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.menus.infra.listener.ProductPriceChangeListener;
 import kitchenpos.menus.tobe.domain.Product;
-import kitchenpos.products.application.FakePurgomalumClient;
+import kitchenpos.clients.FakePurgomalumClient;
+import kitchenpos.menus.tobe.domain.PurgomalumClient;
 import kitchenpos.products.application.InMemoryProductRepository;
-import kitchenpos.products.tobe.domain.*;
+import kitchenpos.products.tobe.domain.ProductPrice;
+import kitchenpos.products.tobe.domain.ProductPriceChangeEvent;
+import kitchenpos.products.tobe.domain.ProductRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
